restructure most of codes. Will add more feartures.
