+++ .env
@@ -1,0 +1,4 @@
+BASE_URL=http://bes.outposter.com.au/api
+GET_USER=/auth/user
+GET_TIMELOG=/timelogs
+POST_CLOCK=/clock
