+++ babel.config.js
@@ -2,6 +2,19 @@ module.exports = function (api) {
   api.cache(true);
   return {
     presets: ["babel-preset-expo"],
-    plugins: ["nativewind/babel", "react-native-reanimated/plugin"],
+    plugins: [
+      "nativewind/babel",
+      "react-native-reanimated/plugin",
+      [
+        "module:react-native-dotenv",
+        {
+          path: ".env",
+          blacklist: null,
+          whitelist: null,
+          safe: false,
+          allowUndefined: true,
+        },
+      ],
+    ],
   };
 };
