+++ package.json
@@ -36,6 +36,7 @@
     "react-icons": "^4.10.1",
     "react-native": "0.71.8",
     "react-native-clock-analog": "^0.0.3",
+    "react-native-dotenv": "^3.4.9",
     "react-native-gesture-handler": "~2.9.0",
     "react-native-paper": "^5.10.6",
     "react-native-reanimated": "^2.14.4",
