+++ src/components/profile/darkMode/DarkModeSwitch.js
@@ -1,7 +1,7 @@
 import { StatusBar } from "expo-status-bar";
 import React, { useContext } from "react";
 import { Text, Switch } from "react-native";
-import ThemeContext from "../store/darkMode/theme-context";
+import ThemeContext from "../../../store/darkMode/theme-context";
 
 const DarkModeSwitch = () => {
   const { themeIs, setToggleTheme } = useContext(ThemeContext);
@@ -12,7 +12,7 @@ const DarkModeSwitch = () => {
       <Switch
         trackColor={{ false: "#2B6673", true: "#11292e" }}
         thumbColor="#E5E4E2"
-        className="border-5 absolute top-3 right-2"
+        className="border-5 absolute top-10 right-2"
         value={themeIs === "dark"}
         onChange={() => setToggleTheme(themeIs === "dark" ? "light" : "dark")}
       />
