+++ src/store/darkMode/theme-context.js
@@ -1,6 +1,6 @@
-import React from "react";
+import { createContext } from "react";
 
-const ThemeContext = React.createContext({
+const ThemeContext = createContext({
   themeIs: null,
   setToggleTheme: () => {},
 });
